initial commit
This is my main reboot script.  This script is not ready and will
require modification in order for it to work on any server other than
mine.

This is provided mostly for reference.
This is an example of a script to use when automating maintenance kicked
off by scheduled SE server reboot.

it is messy and hard to read and meant for those familiar with
powershell.
provided here as a request.

I am the original author of this script in its entirety and freely
authorize any modifications.  I just ask that you give credit to Tostito
as the original author.